Move function in lib/
